+++ src/components/spartan/STemplateHeaderTable/STemplateHeaderTable.vue
@@ -1,23 +1,40 @@
 <script setup lang="ts">
-import type { TTableWithSearchProps } from './types';
+import type { TTemplateHeaderTableProps, TTemplateHeaderTableEmits } from './types';
+import { translator } from '@/helpers';
+import { ref } from 'vue';
+import { MagnifyingGlassIcon } from '@heroicons/vue/20/solid';
+import { SInput, SButton } from '@spartan';
 
-defineProps<TTableWithSearchProps>();
+const emit = defineEmits<TTemplateHeaderTableEmits>();
+
+defineProps<TTemplateHeaderTableProps>();
+
+const { t } = translator('templateHeaderTable');
+
+const query = ref('');
+
+const search = () => {
+    emit('search', query.value);
+};
 </script>
 
 <template>
-    <a
-        v-bind="$attrs"
-        class="cursor-pointer font-medium hover:text-gray-700 hover:underline"
-    >
-        <slot />
-    </a>
-<STemplateHeaderTable>
-    <template>
-        <SButton>
-            <template #icon>
-                <MagnifyingGlassIcon />
-            </template>
-        </SButton>
-    </template>
-</STemplateHeaderTable>
+    <div class="flex flex-col gap-4">
+        <div class="flex items-center justify-between gap-4">
+            <h1 class="text-2xl font-bold text-gray-900">{{ title }}</h1>
+
+            <div class="flex -space-x-px">
+                <SInput
+                    v-model="query"
+                    placeholder="Find your product..."
+                    rounded="left"
+                    required
+                    @keyup.enter="search"
+                />
+                <SButton rounded="right" :icon="MagnifyingGlassIcon" @click="search">{{ t('search') }}</SButton>
+            </div>
+        </div>
+
+        <slot name="table" />
+    </div>
 </template>