+++ src/components/spartan/STemplateHeaderTable/types.ts
@@ -1,4 +1,7 @@
-export type TTableWithSearchProps = {
-    data: Record<string, any>[];
-    search: string;
+export type TTemplateHeaderTableProps = {
+    title: string;
+};
+
+export type TTemplateHeaderTableEmits = {
+    (e: 'search', query: string): void;
 };